+++ test/osm.test.js
@@ -10,7 +10,7 @@ describe("osm (xml)", function () {
     var xml = "<osm></osm>";
     xml = (new DOMParser()).parseFromString(xml, 'text/xml');
 
-    expect(osmtogeojson.toGeojson(xml)).to.eql({
+    expect(osmtogeojson(xml, {flatProperties: false})).to.eql({
       type: 'FeatureCollection',
       features: []
     });
@@ -42,7 +42,7 @@ describe("osm (xml)", function () {
       ]
     };
 
-    expect(osmtogeojson.toGeojson(xml)).to.eql(geojson);
+    expect(osmtogeojson(xml, {flatProperties: false})).to.eql(geojson);
   });
 
   it('way', function() {
@@ -75,7 +75,7 @@ describe("osm (xml)", function () {
       ]
     };
 
-    expect(osmtogeojson.toGeojson(xml)).to.eql(geojson);
+    expect(osmtogeojson(xml, {flatProperties: false})).to.eql(geojson);
   });
 
   it('relation', function() {
@@ -122,7 +122,7 @@ describe("osm (xml)", function () {
       ]
     };
 
-    expect(osmtogeojson.toGeojson(xml)).to.eql(geojson);
+    expect(osmtogeojson(xml, {flatProperties: false})).to.eql(geojson);
   });
 
 
@@ -162,7 +162,7 @@ describe("osm (json)", function () {
         }
       ]
     };
-    var result = osmtogeojson.toGeojson(json);
+    var result = osmtogeojson(json, {flatProperties: false});
     expect(result).to.eql(geojson);
   });
   it("way", function () {
@@ -218,7 +218,7 @@ describe("osm (json)", function () {
         }
       ]
     };
-    var result = osmtogeojson.toGeojson(json);
+    var result = osmtogeojson(json, {flatProperties: false});
     expect(result).to.eql(geojson);
   });
   it("polygon", function () {
@@ -283,7 +283,7 @@ describe("osm (json)", function () {
         }
       ]
     };
-    var result = osmtogeojson.toGeojson(json);
+    var result = osmtogeojson(json, {flatProperties: false});
     expect(result).to.eql(geojson);
   });
   it("simple multipolygon", function () {
@@ -400,7 +400,7 @@ describe("osm (json)", function () {
         }
       ]
     };
-    var result = osmtogeojson.toGeojson(json);
+    var result = osmtogeojson(json, {flatProperties: false});
     expect(result).to.eql(geojson);
     // invalid simple multipolygon (no outer way)
     json = {
@@ -429,7 +429,7 @@ describe("osm (json)", function () {
       features: [
       ]
     };
-    result = osmtogeojson.toGeojson(json);
+    result = osmtogeojson(json, {flatProperties: false});
     expect(result).to.eql(geojson);
   });
   it("multipolygon", function () {
@@ -688,12 +688,121 @@ describe("osm (json)", function () {
         }
       ]
     };
-    var result = osmtogeojson.toGeojson(json);
+    var result = osmtogeojson(json, {flatProperties: false});
     expect(result).to.eql(geojson);
     // handle role-less members as outer ways
     json.elements[0].members[3].role = "";
     geojson.features[2].properties.relations[0].role = "";
-    result = osmtogeojson.toGeojson(json);
+    result = osmtogeojson(json, {flatProperties: false});
+    expect(result).to.eql(geojson);
+  });
+  it("route relation", function () {
+    var json, geojson;
+    // valid route relation
+    json = {
+      elements: [
+        {
+          type:    "relation",
+          id:      1,
+          tags:    {"type":"route"},
+          members: [
+            {
+              type: "way",
+              ref:  2,
+              role: "forward"
+            },
+            {
+              type: "way",
+              ref:  3,
+              role: "backward"
+            },
+            {
+              type: "way",
+              ref:  4,
+              role: "forward"
+            }
+          ]
+        },
+        {
+          type:  "way",
+          id:    2,
+          nodes: [4,5]
+        },
+        {
+          type:  "way",
+          id:    3,
+          nodes: [5,6]
+        },
+        {
+          type:  "way",
+          id:    4,
+          nodes: [7,8]
+        },
+        {
+          type: "node",
+          id:   4,
+          lat: -1.0,
+          lon: -1.0
+        },
+        {
+          type: "node",
+          id:   5,
+          lat:  0.0,
+          lon:  0.0
+        },
+        {
+          type: "node",
+          id:   6,
+          lat:  1.0,
+          lon:  1.0
+        },
+        {
+          type: "node",
+          id:   7,
+          lat:  10.0,
+          lon:  10.0
+        },
+        {
+          type: "node",
+          id:   8,
+          lat:  20.0,
+          lon:  20.0
+        }
+      ]
+    };
+    geojson = {
+      type: "FeatureCollection",
+      features: [
+        {
+          type: "Feature",
+          id: "relation/1",
+          properties: {
+            type: "relation",
+            id: 1,
+            tags: {"type":"route"},
+            relations: [],
+            meta: {}
+          },
+          geometry: {
+            type: "MultiLineString",
+            coordinates: [[
+              [-1.0,-1.0],
+              [ 0.0, 0.0],
+              [ 1.0, 1.0]
+            ],[
+              [10.0,10.0],
+              [20.0,20.0]
+            ]]
+          }
+        }
+      ]
+    };
+    var result = osmtogeojson(json, {flatProperties: false});
+    function _sorter(a,b) {
+      return a.length - b.length;
+    }
+    result.features[0].geometry.coordinates.sort(_sorter);
+    geojson.features[0].geometry.coordinates.sort(_sorter);
     expect(result).to.eql(geojson);
   });
   // tags & pois
@@ -789,7 +898,7 @@ describe("osm (json)", function () {
         }
       ]
     };
-    var result = osmtogeojson.toGeojson(json);
+    var result = osmtogeojson(json, {flatProperties: false});
     expect(result).to.eql(geojson);
   });
   // invalid one-node-ways
@@ -811,11 +920,11 @@ describe("osm (json)", function () {
         }
       ]
     };
-    result = osmtogeojson.toGeojson(json);
+    result = osmtogeojson(json, {flatProperties: false});
     expect(result.features).to.have.length(0);
   });
   // invalid empty multipolygon
-  it("empty multipolygon", function () {
+  it("invalid multipolygon: empty", function () {
     var json, result;
     // empty multipolygon
     json = {
@@ -827,9 +936,119 @@ describe("osm (json)", function () {
         }
       ]
     };
-    result = osmtogeojson.toGeojson(json);
+    result = osmtogeojson(json, {flatProperties: false});
     expect(result.features).to.have.length(0);
   });
+  // invalid multipolygon with missing members
+  it("invalid multipolygon: missing members", function () {
+    var json, result;
+    // empty multipolygon
+    json = {
+      elements: [
+        {
+          type: "relation",
+          id:   1,
+          tags: {"type":"multipolygon"},
+          members: [{
+            type: "way",
+            ref:  1,
+            role: "outer"
+          }]
+        }
+      ]
+    };
+    result = osmtogeojson(json, {flatProperties: false});
+    expect(result.features).to.have.length(0);
+  });
+  // invalid multipolygon with empty members
+  it("invalid multipolygon: empty members", function () {
+    var json, result;
+    // empty multipolygon
+    json = {
+      elements: [
+        {
+          type: "relation",
+          id:   1,
+          tags: {"type":"multipolygon"},
+          members: [{
+            type: "way",
+            ref:  1,
+            role: "outer"
+          }]
+        },
+        {
+          type: "way",
+          id: 1
+        }
+      ]
+    };
+    result = osmtogeojson(json, {flatProperties: false});
+    expect(result.features).to.have.length(0);
+  });
+
+  // invalid empty route
+  it("invalid route: empty", function () {
+    var json, result;
+    // empty multipolygon
+    json = {
+      elements: [
+        {
+          type: "relation",
+          id:   1,
+          tags: {"type":"route"}
+        }
+      ]
+    };
+    result = osmtogeojson(json, {flatProperties: false});
+    expect(result.features).to.have.length(0);
+  });
+  // invalid route with missing members
+  it("invalid route: missing members", function () {
+    var json, result;
+    // empty multipolygon
+    json = {
+      elements: [
+        {
+          type: "relation",
+          id:   1,
+          tags: {"type":"route"},
+          members: [{
+            type: "way",
+            ref:  1,
+            role: "forward"
+          }]
+        }
+      ]
+    };
+    result = osmtogeojson(json, {flatProperties: false});
+    expect(result.features).to.have.length(0);
+  });
+  // invalid route with empty members
+  it("invalid route: empty members", function () {
+    var json, result;
+    // empty multipolygon
+    json = {
+      elements: [
+        {
+          type: "relation",
+          id:   1,
+          tags: {"type":"route"},
+          members: [{
+            type: "way",
+            ref:  1,
+            role: "forward"
+          }]
+        },
+        {
+          type: "way",
+          id: 1
+        }
+      ]
+    };
+    result = osmtogeojson(json, {flatProperties: false});
+    expect(result.features).to.have.length(0);
+  });
+
   // relations
   it("relations and id-spaces", function () {
     var json, geojson;
@@ -987,7 +1206,7 @@ describe("osm (json)", function () {
         }
       ]
     };
-    var result = osmtogeojson.toGeojson(json);
+    var result = osmtogeojson(json, {flatProperties: false});
     expect(result).to.eql(geojson);
   });
   // meta info // todo +lines, +polygons
@@ -1035,7 +1254,7 @@ describe("osm (json)", function () {
         }
       ]
     };
-    result = osmtogeojson.toGeojson(json);
+    result = osmtogeojson(json, {flatProperties: false});
     expect(result).to.eql(geojson);
     // ways and relsvar json, geojson;
     json = {
@@ -1078,7 +1297,7 @@ describe("osm (json)", function () {
         }
       ]
     };
-    result = osmtogeojson.toGeojson(json);
+    result = osmtogeojson(json, {flatProperties: false});
     expect(result.features).to.have.length(4);
     expect(result.features[0].properties.meta).to.have.property("user");
     expect(result.features[1].properties.meta).to.have.property("user");
@@ -1163,7 +1382,7 @@ describe("osm (json)", function () {
         }
       ]
     };
-    var result = osmtogeojson.toGeojson(json);
+    var result = osmtogeojson(json, {flatProperties: false});
     expect(result.features).to.have.length(2);
     expect(result.features[0].properties.id).to.eql(1);
     expect(result.features[1].properties.id).to.eql(2);
@@ -1249,7 +1468,7 @@ describe("osm (json)", function () {
         }
       ]
     };
-    result = osmtogeojson.toGeojson(json);
+    result = osmtogeojson(json, {flatProperties: false});
     expect(result.features).to.have.length(1);
     expect(result.features[0].properties.id).to.equal(1);
     expect(result.features[0].geometry.type).to.equal("Polygon");
@@ -1329,7 +1548,7 @@ describe("osm (json)", function () {
         }
       ]
     };
-    result = osmtogeojson.toGeojson(json);
+    result = osmtogeojson(json, {flatProperties: false});
     expect(result.features).to.have.length(1);
     expect(result.features[0].properties.id).to.equal(2);
     expect(result.features[0].geometry.type).to.equal("Polygon");
@@ -1397,7 +1616,7 @@ describe("osm (json)", function () {
         }
       ]
     };
-    result = osmtogeojson.toGeojson(json);
+    result = osmtogeojson(json, {flatProperties: false});
     expect(result.features).to.have.length(1);
     expect(result.features[0].properties.id).to.equal(1);
     expect(result.features[0].geometry.type).to.equal("Polygon");
@@ -1511,7 +1730,7 @@ describe("osm (json)", function () {
         }
       ]
     };
-    result = osmtogeojson.toGeojson(json);
+    result = osmtogeojson(json, {flatProperties: false});
     expect(result.features).to.have.length(1);
     expect(result.features[0].properties.id).to.equal(1);
     expect(result.features[0].geometry.type).to.equal("Polygon");
@@ -1576,7 +1795,7 @@ describe("osm (json)", function () {
         }
       ]
     };
-    result = osmtogeojson.toGeojson(json);
+    result = osmtogeojson(json, {flatProperties: false});
     expect(result.features).to.have.length(1);
     expect(result.features[0].properties.id).to.equal(1);
     expect(result.features[0].geometry.type).to.equal("Polygon");
@@ -1639,7 +1858,7 @@ describe("osm (json)", function () {
         }
       ]
     };
-    result = osmtogeojson.toGeojson(json);
+    result = osmtogeojson(json, {flatProperties: false});
     expect(result.features).to.have.length(1);
     expect(result.features[0].properties.id).to.equal(1);
     expect(result.features[0].geometry.type).to.equal("Polygon");
@@ -1658,7 +1877,7 @@ describe("osm (json)", function () {
         }
       ]
     };
-    var result = osmtogeojson.toGeojson(json);
+    var result = osmtogeojson(json, {flatProperties: false});
     expect(result.features).to.have.length(0);
   });
 
@@ -1691,7 +1910,7 @@ describe("defaults", function() {
         }
       ]
     };
-    var result = osmtogeojson.toGeojson(json);
+    var result = osmtogeojson(json, {flatProperties: false});
     expect(result.features).to.have.length(2);
     expect(result.features[0].geometry.type).to.equal("LineString");
     expect(result.features[1].geometry.type).to.equal("Point");
@@ -1753,7 +1972,7 @@ describe("defaults", function() {
             '<node id="4295668181" version="1" changeset="23123" lat="46.4923235" lon="11.2752747" user="tyrTester06" uid="1178" visible="true" timestamp="2013-05-14T10:33:04Z"/>'+
           '</osm>'
     xml = (new DOMParser()).parseFromString(xml, 'text/xml');
-    var result = osmtogeojson.toGeojson(xml);
+    var result = osmtogeojson(xml, {flatProperties: false});
     expect(result.features).to.have.length(8);
   });
   // polygon detection
@@ -1795,7 +2014,7 @@ describe("defaults", function() {
         }
       ]
     };
-    result = osmtogeojson.toGeojson(json);
+    result = osmtogeojson(json, {flatProperties: false});
     expect(result.features).to.have.length(2);
     expect(result.features[0].geometry.type).to.equal("Polygon");
     expect(result.features[1].geometry.type).to.equal("LineString");
@@ -1831,7 +2050,7 @@ describe("defaults", function() {
         }
       ]
     };
-    result = osmtogeojson.toGeojson(json);
+    result = osmtogeojson(json, {flatProperties: false});
     expect(result.features).to.have.length(1);
     expect(result.features[0].geometry.type).to.equal("LineString");
   });
@@ -1858,7 +2077,10 @@ describe("options", function () {
       foo: "bar",
       user: "johndoe"
     };
-    var result = osmtogeojson.toGeojson(json, {flatProperties: true});
+    var result = osmtogeojson(json, {flatProperties: true});
+    expect(result.features[0].properties).to.eql(geojson_properties);
+    // check default
+    result = osmtogeojson(json);
     expect(result.features[0].properties).to.eql(geojson_properties);
   });
   // interesting objects
@@ -1889,7 +2111,7 @@ describe("options", function () {
         }
       ]
     };
-    var result = osmtogeojson.toGeojson(json, {uninterestingTags: {foo:true}});
+    var result = osmtogeojson(json, {flatProperties: false, uninterestingTags: {foo:true}});
     expect(result.features).to.have.length(2);
     expect(result.features[1].properties.id).to.eql(3);
   });
@@ -1919,7 +2141,7 @@ describe("options", function () {
         }
       ]
     };
-    var result = osmtogeojson.toGeojson(json, {uninterestingTags: function(tags, ignore_tags) {
+    var result = osmtogeojson(json, {flatProperties: false, uninterestingTags: function(tags, ignore_tags) {
       return tags["tag"] != "1";
     }});
     expect(result.features).to.have.length(2);
@@ -1990,7 +2212,7 @@ describe("options", function () {
         }
       ]
     };
-    result = osmtogeojson.toGeojson(json, {
+    result = osmtogeojson(json, {
       polygonFeatures: {
         "is_polygon_key": true,
         "is_polygon_key_value": {
@@ -2046,7 +2268,7 @@ describe("options", function () {
         }
       ]
     };
-    var result = osmtogeojson.toGeojson(json, {polygonFeatures: function(tags) {
+    var result = osmtogeojson(json, {polygonFeatures: function(tags) {
       return tags["tag"] == "1";
     }});
     expect(result.features).to.have.length(2);
@@ -2186,7 +2408,7 @@ describe("tainted data", function () {
         }
       ]
     };
-    var result = osmtogeojson.toGeojson(json);
+    var result = osmtogeojson(json, {flatProperties: false});
     expect(result).to.eql(geojson);
   });
   // ignore missing node coordinates
@@ -2200,7 +2422,7 @@ describe("tainted data", function () {
         }
       ]
     };
-    result = osmtogeojson.toGeojson(json);
+    result = osmtogeojson(json, {flatProperties: false});
     expect(result.features).to.have.length(0);
     json = {
       elements: [
@@ -2210,7 +2432,7 @@ describe("tainted data", function () {
         }
       ]
     };
-    result = osmtogeojson.toGeojson(json);
+    result = osmtogeojson(json, {flatProperties: false});
     expect(result.features).to.have.length(0);
     json = {
       elements: [
@@ -2220,7 +2442,7 @@ describe("tainted data", function () {
         }
       ]
     };
-    result = osmtogeojson.toGeojson(json);
+    result = osmtogeojson(json, {flatProperties: false});
     expect(result.features).to.have.length(0);
   });
   // tainted way
@@ -2247,7 +2469,7 @@ describe("tainted data", function () {
         }
       ]
     };
-    var result = osmtogeojson.toGeojson(json);
+    var result = osmtogeojson(json, {flatProperties: false});
     expect(result.features).to.have.length(1);
     expect(result.features[0].properties.id).to.equal(1);
     expect(result.features[0].geometry.coordinates).to.eql([[0.0,0.0],[1.0,1.0]]);
@@ -2305,7 +2527,7 @@ describe("tainted data", function () {
         }
       ]
     };
-    result = osmtogeojson.toGeojson(json);
+    result = osmtogeojson(json, {flatProperties: false});
     expect(result.features).to.have.length(1);
     expect(result.features[0].properties.id).to.equal(2);
     expect(result.features[0].properties.tainted).to.equal(true);
@@ -2331,7 +2553,7 @@ describe("tainted data", function () {
         }
       ]
     };
-    result = osmtogeojson.toGeojson(json);
+    result = osmtogeojson(json, {flatProperties: false});
     expect(result.features).to.have.length(0);
     // missing node
     json = {
@@ -2373,7 +2595,7 @@ describe("tainted data", function () {
         }
       ]
     };
-    result = osmtogeojson.toGeojson(json);
+    result = osmtogeojson(json, {flatProperties: false});
     expect(result.features).to.have.length(1);
     expect(result.features[0].properties.id).to.equal(2);
     expect(result.features[0].properties.tainted).to.equal(true);
@@ -2426,7 +2648,7 @@ describe("tainted data", function () {
         }
       ]
     };
-    result = osmtogeojson.toGeojson(json);
+    result = osmtogeojson(json, {flatProperties: false});
     expect(result.features).to.have.length(1);
     expect(result.features[0].properties.id).to.equal(1);
     expect(result.features[0].properties.tainted).to.equal(true);
@@ -2480,7 +2702,7 @@ describe("tainted data", function () {
         }
       ]
     };
-    result = osmtogeojson.toGeojson(json);
+    result = osmtogeojson(json, {flatProperties: false});
     expect(result.features).to.have.length(1);
     expect(result.features[0].properties.id).to.equal(1);
     expect(result.features[0].properties.tainted).to.equal(true);
@@ -2520,7 +2742,7 @@ describe("tainted data", function () {
         }
       ]
     };
-    result = osmtogeojson.toGeojson(json);
+    result = osmtogeojson(json, {flatProperties: false});
     expect(result.features).to.have.length(0);
     // no outer ring
     json = {
@@ -2562,7 +2784,7 @@ describe("tainted data", function () {
         }
       ]
     };
-    result = osmtogeojson.toGeojson(json);
+    result = osmtogeojson(json, {flatProperties: false});
     // expected behaviour: do not return a degenerate (multi)polygon.
     // this could in principle return just the way that is now sort of unused
     // but just as with an (untagged) node of a one-node-way we're going to
@@ -2607,7 +2829,7 @@ describe("tainted data", function () {
         }
       ]
     };
-    result = osmtogeojson.toGeojson(json);
+    result = osmtogeojson(json, {flatProperties: false});
     expect(result.features).to.have.length(0);
   });
 
@@ -2644,7 +2866,7 @@ describe("other", function () {
       ]
     };
     json_before = JSON.stringify(json);
-    osmtogeojson.toGeojson(json);
+    osmtogeojson(json, {flatProperties: false});
     json_after = JSON.stringify(json);
     expect(json_after).to.equal(json_before);
   });
@@ -2660,7 +2882,7 @@ describe("overpass geometry types", function () {
     xml = "<osm><way id='1'><center lat='1.234' lon='4.321' /></way></osm>";
     xml = (new DOMParser()).parseFromString(xml, 'text/xml');
 
-    geojson = osmtogeojson.toGeojson(xml);
+    geojson = osmtogeojson(xml, {flatProperties: false});
 
     expect(geojson.features.length).to.eql(1);
     expect(geojson.features[0].id).to.eql("way/1");
@@ -2672,7 +2894,7 @@ describe("overpass geometry types", function () {
     xml = "<osm><relation id='1'><center lat='1.234' lon='4.321' /></relation></osm>";
     xml = (new DOMParser()).parseFromString(xml, 'text/xml');
 
-    geojson = osmtogeojson.toGeojson(xml);
+    geojson = osmtogeojson(xml, {flatProperties: false});
 
     expect(geojson.features.length).to.eql(1);
     expect(geojson.features[0].id).to.eql("relation/1");
@@ -2696,7 +2918,7 @@ describe("overpass geometry types", function () {
         }
       ]
     };
-    geojson = osmtogeojson.toGeojson(json);
+    geojson = osmtogeojson(json, {flatProperties: false});
 
     expect(geojson.features.length).to.eql(1);
     expect(geojson.features[0].id).to.eql("way/1");
@@ -2717,7 +2939,7 @@ describe("overpass geometry types", function () {
         }
       ]
     };
-    geojson = osmtogeojson.toGeojson(json);
+    geojson = osmtogeojson(json, {flatProperties: false});
 
     expect(geojson.features.length).to.eql(1);
     expect(geojson.features[0].id).to.eql("relation/1");
@@ -2734,7 +2956,7 @@ describe("overpass geometry types", function () {
     xml = "<osm><way id='1'><bounds minlat='1.234' minlon='4.321' maxlat='2.234' maxlon='5.321'/></way></osm>";
     xml = (new DOMParser()).parseFromString(xml, 'text/xml');
 
-    geojson = osmtogeojson.toGeojson(xml);
+    geojson = osmtogeojson(xml, {flatProperties: false});
 
     expect(geojson.features.length).to.eql(1);
     expect(geojson.features[0].id).to.eql("way/1");
@@ -2745,7 +2967,7 @@ describe("overpass geometry types", function () {
     xml = "<osm><relation id='1'><bounds minlat='1.234' minlon='4.321' maxlat='2.234' maxlon='5.321'/></relation></osm>";
     xml = (new DOMParser()).parseFromString(xml, 'text/xml');
 
-    geojson = osmtogeojson.toGeojson(xml);
+    geojson = osmtogeojson(xml, {flatProperties: false});
 
     expect(geojson.features.length).to.eql(1);
     expect(geojson.features[0].id).to.eql("relation/1");
@@ -2770,7 +2992,7 @@ describe("overpass geometry types", function () {
         }
       ]
     };
-    geojson = osmtogeojson.toGeojson(json);
+    geojson = osmtogeojson(json, {flatProperties: false});
 
     expect(geojson.features.length).to.eql(1);
     expect(geojson.features[0].id).to.eql("way/1");
@@ -2792,7 +3014,7 @@ describe("overpass geometry types", function () {
         }
       ]
     };
-    geojson = osmtogeojson.toGeojson(json);
+    geojson = osmtogeojson(json, {flatProperties: false});
 
     expect(geojson.features.length).to.eql(1);
     expect(geojson.features[0].id).to.eql("relation/1");
@@ -2815,7 +3037,7 @@ describe("overpass geometry types", function () {
         + "</way></osm>";
     xml = (new DOMParser()).parseFromString(xml, 'text/xml');
 
-    geojson = osmtogeojson.toGeojson(xml);
+    geojson = osmtogeojson(xml, {flatProperties: false});
 
     expect(geojson.features.length).to.eql(1);
     expect(geojson.features[0].id).to.eql("way/1");
@@ -2833,7 +3055,7 @@ describe("overpass geometry types", function () {
         + "</way></osm>";
     xml = (new DOMParser()).parseFromString(xml, 'text/xml');
 
-    geojson = osmtogeojson.toGeojson(xml);
+    geojson = osmtogeojson(xml, {flatProperties: false});
 
     expect(geojson.features.length).to.eql(1);
     expect(geojson.features[0].id).to.eql("way/1");
@@ -2862,7 +3084,7 @@ describe("overpass geometry types", function () {
         + "</relation></osm>";
     xml = (new DOMParser()).parseFromString(xml, 'text/xml');
 
-    geojson = osmtogeojson.toGeojson(xml);
+    geojson = osmtogeojson(xml, {flatProperties: false});
 
     expect(geojson.features.length).to.eql(2);
     expect(geojson.features[0].id).to.eql("relation/1");
@@ -2909,7 +3131,7 @@ describe("overpass geometry types", function () {
         + "</osm>";
     xml = (new DOMParser()).parseFromString(xml, 'text/xml');
 
-    geojson = osmtogeojson.toGeojson(xml);
+    geojson = osmtogeojson(xml, {flatProperties: false});
 
     expect(geojson.features.length).to.eql(2);
     expect(geojson.features[0].geometry.type).to.eql("Polygon");
@@ -2952,7 +3174,7 @@ describe("overpass geometry types", function () {
         }
       ]
     };
-    geojson = osmtogeojson.toGeojson(json);
+    geojson = osmtogeojson(json, {flatProperties: false});
 
     expect(geojson.features.length).to.eql(1);
     expect(geojson.features[0].id).to.eql("way/1");
@@ -2983,7 +3205,7 @@ describe("overpass geometry types", function () {
         }
       ]
     };
-    geojson = osmtogeojson.toGeojson(json);
+    geojson = osmtogeojson(json, {flatProperties: false});
 
     expect(geojson.features.length).to.eql(1);
     expect(geojson.features[0].id).to.eql("way/1");
@@ -3040,7 +3262,7 @@ describe("overpass geometry types", function () {
         }
       ]
     };
-    geojson = osmtogeojson.toGeojson(json);
+    geojson = osmtogeojson(json, {flatProperties: false});
 
     expect(geojson.features.length).to.eql(2);
     expect(geojson.features[0].id).to.eql("relation/1");
@@ -3129,7 +3351,7 @@ describe("overpass geometry types", function () {
         }
       ]
     };
-    geojson = osmtogeojson.toGeojson(json);
+    geojson = osmtogeojson(json, {flatProperties: false});
 
     expect(geojson.features.length).to.eql(2);
     expect(geojson.features[0].geometry.type).to.eql("Polygon");
@@ -3155,7 +3377,7 @@ describe("overpass geometry types", function () {
         + "</node></osm>";
     xml = (new DOMParser()).parseFromString(xml, 'text/xml');
 
-    geojson = osmtogeojson.toGeojson(xml);
+    geojson = osmtogeojson(xml, {flatProperties: false});
 
     expect(geojson.features.length).to.eql(2);
     expect(geojson.features[0].id).to.eql("way/1");
@@ -3193,7 +3415,7 @@ describe("overpass geometry types", function () {
         }
       ]
     };
-    geojson = osmtogeojson.toGeojson(json);
+    geojson = osmtogeojson(json, {flatProperties: false});
 
     expect(geojson.features.length).to.eql(2);
     expect(geojson.features[0].id).to.eql("way/1");
@@ -3214,7 +3436,7 @@ describe("overpass geometry types", function () {
         + "</way></osm>";
     xml = (new DOMParser()).parseFromString(xml, 'text/xml');
 
-    geojson = osmtogeojson.toGeojson(xml);
+    geojson = osmtogeojson(xml, {flatProperties: false});
 
     expect(geojson.features.length).to.eql(1);
     expect(geojson.features[0].id).to.eql("way/1");
@@ -3231,7 +3453,7 @@ describe("overpass geometry types", function () {
         + "</way></osm>";
     xml = (new DOMParser()).parseFromString(xml, 'text/xml');
 
-    geojson = osmtogeojson.toGeojson(xml);
+    geojson = osmtogeojson(xml, {flatProperties: false});
 
     expect(geojson.features.length).to.eql(1);
     expect(geojson.features[0].id).to.eql("way/1");
@@ -3265,7 +3487,7 @@ describe("overpass geometry types", function () {
         + "</osm>";
     xml = (new DOMParser()).parseFromString(xml, 'text/xml');
 
-    geojson = osmtogeojson.toGeojson(xml);
+    geojson = osmtogeojson(xml, {flatProperties: false});
 
     expect(geojson.features.length).to.eql(2);
     expect(geojson.features[0].id).to.eql("way/1");
@@ -3303,7 +3525,7 @@ describe("overpass geometry types", function () {
         }
       ]
     };
-    geojson = osmtogeojson.toGeojson(json);
+    geojson = osmtogeojson(json, {flatProperties: false});
 
     expect(geojson.features.length).to.eql(1);
     expect(geojson.features[0].id).to.eql("way/1");
@@ -3326,7 +3548,7 @@ describe("overpass geometry types", function () {
         }
       ]
     };
-    geojson = osmtogeojson.toGeojson(json);
+    geojson = osmtogeojson(json, {flatProperties: false});
 
     expect(geojson.features.length).to.eql(1);
     expect(geojson.features[0].id).to.eql("way/1");
@@ -3387,7 +3609,7 @@ describe("overpass geometry types", function () {
         }
       ]
     };
-    geojson = osmtogeojson.toGeojson(json);
+    geojson = osmtogeojson(json, {flatProperties: false});
 
     expect(geojson.features.length).to.eql(2);
     expect(geojson.features[0].id).to.eql("way/1");
@@ -3458,7 +3680,7 @@ describe("overpass geometry types", function () {
         }
       ]
     };
-    geojson = osmtogeojson.toGeojson(json);
+    geojson = osmtogeojson(json, {flatProperties: false});
 
     expect(geojson.features[0].id).to.eql("relation/1");
     expect(geojson.features[0].geometry.type).to.eql("Polygon");
@@ -3491,7 +3713,7 @@ describe("overpass geometry types", function () {
         + "</osm>";
     xml = (new DOMParser()).parseFromString(xml, 'text/xml');
 
-    geojson = osmtogeojson.toGeojson(xml);
+    geojson = osmtogeojson(xml, {flatProperties: false});
 
     expect(geojson.features[0].id).to.eql("relation/1");
     expect(geojson.features[0].geometry.type).to.eql("Polygon");
@@ -3533,7 +3755,7 @@ describe("duplicate elements", function () {
         }
       ]
     };
-    geojson = osmtogeojson.toGeojson(json);
+    geojson = osmtogeojson(json, {flatProperties: false});
 
     expect(geojson.features.length).to.eql(1);
     expect(geojson.features[0].id).to.eql("node/1");
@@ -3572,7 +3794,7 @@ describe("duplicate elements", function () {
         }
       ]
     };
-    geojson = osmtogeojson.toGeojson(json);
+    geojson = osmtogeojson(json, {flatProperties: false});
 
     expect(geojson.features.length).to.eql(1);
     expect(geojson.features[0].id).to.eql("node/1");
@@ -3615,7 +3837,7 @@ describe("duplicate elements", function () {
         }
       ]
     };
-    geojson = osmtogeojson.toGeojson(json);
+    geojson = osmtogeojson(json, {flatProperties: false});
 
     expect(geojson.features.length).to.eql(1);
     expect(geojson.features[0].id).to.eql("way/1");
@@ -3669,7 +3891,7 @@ describe("duplicate elements", function () {
         }
       ]
     };
-    geojson = osmtogeojson.toGeojson(json);
+    geojson = osmtogeojson(json, {flatProperties: false});
 
     expect(geojson.features.length).to.eql(1);
     expect(geojson.features[0].id).to.eql("way/1");
@@ -3736,13 +3958,75 @@ describe("duplicate elements", function () {
         }
       ]
     };
-    geojson = osmtogeojson.toGeojson(json);
+    geojson = osmtogeojson(json, {flatProperties: false});
 
     expect(geojson.features.length).to.eql(2);
     expect(geojson.features[0].id).to.eql("relation/1");
     expect(geojson.features[1].id).to.eql("way/2");
     expect(geojson.features[0].properties.meta.version).to.eql(2);
     expect(geojson.features[0].properties.tags.foo).to.eql("2");
+  });
+
+  it("relation, additional skeleton ways", function () {
+    var json, geojson;
+
+    // do not include full geometry nd's as node in output
+    json = {
+      elements: [
+        {
+          type: "relation",
+          id: 1,
+          version: 2,
+          members: [{type: "way", ref: 1, role: "outer"}, {type: "way", ref: 2, role: "outer"}],
+          tags: {
+            "type": "multipolygon",
+            "foo": "bar"
+          }
+        },
+        {
+          type: "way",
+          id: 1,
+          nodes: [1,2,3],
+          tags: {
+            "asd": "fasd"
+          }
+        },
+        {
+          type: "way",
+          id: 1,
+          nodes: [1,2,3]
+        },
+        {
+          type: "way",
+          id: 2,
+          nodes: [3,1]
+        },
+        {
+          type: "node",
+          id: 1,
+          lat: 1,
+          lon: 1
+        },
+        {
+          type: "node",
+          id: 2,
+          lat: 2,
+          lon: 2
+        },
+        {
+          type: "node",
+          id: 3,
+          lat: 2,
+          lon: 1
+        }
+      ]
+    };
+    geojson = osmtogeojson(json, {flatProperties: false});
+
+    expect(geojson.features.length).to.eql(2);
+    expect(geojson.features[0].id).to.eql("relation/1");
+    expect(geojson.features[1].id).to.eql("way/1");
+    expect(geojson.features[0].geometry.coordinates.length).to.eql(1);
   });
 
   it("custom deduplicator", function () {
@@ -3775,7 +4059,8 @@ describe("duplicate elements", function () {
         }
       ]
     };
-    geojson = osmtogeojson.toGeojson(json, {
+    geojson = osmtogeojson(json, {
+      flatProperties: false,
       deduplicator: function(a,b) {
         return a.version < b.version ? a : b;
       }
@@ -3789,3 +4074,285 @@ describe("duplicate elements", function () {
   });
 
 });
+
+describe('featureCallback', function() {
+
+  it("node", function () {
+    var json, geojson;
+    json = {
+      elements: [
+        {
+          type: "node",
+          id:   1,
+          lat:  2.0,
+          lon:  3.0
+        }
+      ]
+    };
+    geojson = {
+      type: "Feature",
+      id: "node/1",
+      properties: {
+        type: "node",
+        id: 1,
+        tags: {},
+        relations: [],
+        meta: {}
+      },
+      geometry: {
+        type: "Point",
+        coordinates: [3.0,2.0]
+      }
+    };
+    var result = osmtogeojson(json, {flatProperties: false}, function(feature) {
+      expect(feature).to.eql(geojson);
+    });
+    expect(result).to.equal(true);
+  });
+
+  it("way (line)", function () {
+    var json, geojson;
+    json = {
+      elements: [
+        {
+          type:  "way",
+          id:    1,
+          nodes: [2,3]
+        },
+        {
+          type: "node",
+          id:   2,
+          lat:  0.0,
+          lon:  0.0
+        },
+        {
+          type: "node",
+          id:   3,
+          lat:  1.0,
+          lon:  0.0
+        }
+      ]
+    };
+    geojson = {
+      type: "Feature",
+      id: "way/1",
+      properties: {
+        type: "way",
+        id: 1,
+        tags: {},
+        relations: [],
+        meta: {}
+      },
+      geometry: {
+        type: "LineString",
+        coordinates: [[0.0, 0.0], [0.0,1.0]]
+      }
+    };
+    var result = osmtogeojson(json, {flatProperties: false}, function(feature) {
+      expect(feature).to.eql(geojson);
+    });
+    expect(result).to.equal(true);
+  });
+
+  it("way (polygon)", function () {
+    var json, geojson;
+    json = {
+      elements: [
+        {
+          type:  "way",
+          id:    1,
+          nodes: [2,3,4,5,2],
+          tags:  {area: "yes"}
+        },
+        {
+          type: "node",
+          id:   2,
+          lat:  0.0,
+          lon:  0.0
+        },
+        {
+          type: "node",
+          id:   3,
+          lat:  1.0,
+          lon:  0.0
+        },
+        {
+          type: "node",
+          id:   4,
+          lat:  1.0,
+          lon:  1.0
+        },
+        {
+          type: "node",
+          id:   5,
+          lat:  0.0,
+          lon:  1.0
+        }
+      ]
+    };
+    geojson = {
+      type: "Feature",
+      id: "way/1",
+      properties: {
+        type: "way",
+        id: 1,
+        tags: {area: "yes"},
+        relations: [],
+        meta: {}
+      },
+      geometry: {
+        type: "Polygon",
+        coordinates: [[
+          [0.0,0.0],
+          [1.0,0.0],
+          [1.0,1.0],
+          [0.0,1.0],
+          [0.0,0.0],
+        ]]
+      }
+    };
+    var result = osmtogeojson(json, {flatProperties: false}, function(feature) {
+      expect(feature).to.eql(geojson);
+    });
+    expect(result).to.equal(true);
+  });
+
+  it("relation (simple multipolygon)", function () {
+    var json, geojson;
+    json = {
+      elements: [
+        {
+          type:    "relation",
+          id:      1,
+          members: [{type: "way", ref: 1, role: "outer"}],
+          tags:    {type: "multipolygon"}
+        },
+        {
+          type:  "way",
+          id:    1,
+          nodes: [2,3,4,5,2],
+          tags:  {name: "foo"}
+        },
+        {
+          type: "node",
+          id:   2,
+          lat:  0.0,
+          lon:  0.0
+        },
+        {
+          type: "node",
+          id:   3,
+          lat:  1.0,
+          lon:  0.0
+        },
+        {
+          type: "node",
+          id:   4,
+          lat:  1.0,
+          lon:  1.0
+        },
+        {
+          type: "node",
+          id:   5,
+          lat:  0.0,
+          lon:  1.0
+        }
+      ]
+    };
+    geojson = {
+      type: "Feature",
+      id: "way/1",
+      properties: {
+        type: "way",
+        id: 1,
+        tags: {name: "foo"},
+        relations: [{rel:1, role:"outer", reltags:{type:"multipolygon"}}],
+        meta: {}
+      },
+      geometry: {
+        type: "Polygon",
+        coordinates: [[
+          [0.0,0.0],
+          [1.0,0.0],
+          [1.0,1.0],
+          [0.0,1.0],
+          [0.0,0.0],
+        ]]
+      }
+    };
+    var result = osmtogeojson(json, {flatProperties: false}, function(feature) {
+      expect(feature).to.eql(geojson);
+    });
+    expect(result).to.equal(true);
+  });
+
+  it("relation (multipolygon)", function () {
+    var json, geojson;
+    json = {
+      elements: [
+        {
+          type:    "relation",
+          id:      1,
+          members: [{type: "way", ref: 1, role: "outer"}],
+          tags:    {type: "multipolygon", name: "foo"}
+        },
+        {
+          type:  "way",
+          id:    1,
+          nodes: [2,3,4,5,2],
+          tags:  {}
+        },
+        {
+          type: "node",
+          id:   2,
+          lat:  0.0,
+          lon:  0.0
+        },
+        {
+          type: "node",
+          id:   3,
+          lat:  1.0,
+          lon:  0.0
+        },
+        {
+          type: "node",
+          id:   4,
+          lat:  1.0,
+          lon:  1.0
+        },
+        {
+          type: "node",
+          id:   5,
+          lat:  0.0,
+          lon:  1.0
+        }
+      ]
+    };
+    geojson = {
+      type: "Feature",
+      id: "relation/1",
+      properties: {
+        type: "relation",
+        id: 1,
+        tags: {type: "multipolygon", name: "foo"},
+        relations: [],
+        meta: {}
+      },
+      geometry: {
+        type: "Polygon",
+        coordinates: [[
+          [0.0,0.0],
+          [1.0,0.0],
+          [1.0,1.0],
+          [0.0,1.0],
+          [0.0,0.0],
+        ]]
+      }
+    };
+    var result = osmtogeojson(json, {flatProperties: false}, function(feature) {
+      expect(feature).to.eql(geojson);
+    });
+    expect(result).to.equal(true);
+  });
+
+})